Enum pact_matching::Mismatch

pub enum Mismatch {
    MethodMismatch {
        expected: String,
        actual: String,
    },
    PathMismatch {
        expected: String,
        actual: String,
    },
    StatusMismatch {
        expected: u16,
        actual: u16,
    },
    QueryMismatch {
        parameter: String,
        expected: String,
        actual: String,
        mismatch: String,
    },
    HeaderMismatch {
        key: String,
        expected: String,
        actual: String,
        mismatch: String,
    },
    BodyTypeMismatch {
        expected: String,
        actual: String,
    },
    BodyMismatch {
        path: String,
        expected: Option<String>,
        actual: Option<String>,
        mismatch: String,
    },
}

Enum that defines the different types of mismatches that can occur.

Variants

MethodMismatch

Request Method mismatch

Fields

expected

Expected request method

actual

Actual request method

PathMismatch

Request Path mismatch

Fields

expected

expected request path

actual

actual request path

StatusMismatch

Response status mismatch

Fields

expected

expected response status

actual

actual response status

QueryMismatch

Request query mismatch

Fields

parameter

query parameter name

expected

expected value

actual

actual value

mismatch

description of the mismatch

HeaderMismatch

Header mismatch

Fields

key

header key

expected

expected value

actual

actual value

mismatch

description of the mismatch

BodyTypeMismatch

Mismatch in the content type of the body

Fields

expected

expected content type of the body

actual

actual content type of the body

BodyMismatch

Body element mismatch

Fields

path

path expression to where the mismatch occured

expected

expected value

actual

actual value

mismatch

description of the mismatch

Methods

impl Mismatch

fn to_json(&self) -> Json

Converts the mismatch to a Json struct.

fn mismatch_type(&self) -> String

Returns the type of the mismatch as a string