Module pact_matching::models

The models module provides all the structures required to model a Pact.

Structs

Consumer
Struct that defines the consumer of the pact.

Interaction
Struct that defines an interaction (request and response pair)

Pact
Struct that represents a pact between the consumer and provider of a service.

Provider
Struct that defines a provider of a pact.

Request
Struct that defines the request.

Response
Struct that defines the response.

Enums

OptionalBody
Enum that defines the four main states that a body of a request and response can be in a pact file.

PactSpecification
Enum defining the pact specification versions supported by the library

Traits

HttpPart
Trait to specify an HTTP part of a message. It encapsulates the shared parts of a request and response.

Functions

parse_query_string
Parses a query string into an optional map. The query parameter name will be mapped to a list of values. Where the query parameter is repeated, the order of the values will be preserved.